+++ API/Controllers/UsersController.cs
@@ -1,4 +1,5 @@
 ﻿using API.Models;
+using API.Utilities;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
@@ -17,10 +18,30 @@ namespace API.Controllers
             this.context = context;
         }
 
+        private int RequestUserId
+        {
+            get { return int.Parse(Request.Headers["userId"]); }
+        }
+
+        private bool Authenticate()
+        {
+            if (Authentifier.context == null)
+                Authentifier.context = context;
+
+            string id = Request.Headers["userId"];
+            string password = Request.Headers["userPass"];
+            
+
+            return Authentifier.Authenticate(id, password);
+        }
+
         // GET api/users
         [HttpGet]
         public IEnumerable<User> Get()
         {
+            if (!Authenticate())
+                return null;
+
             return context.Users;
         }
 
@@ -28,7 +49,12 @@ namespace API.Controllers
         [HttpGet("{id}")]
         public User Get(int id)
         {
-            return context.Users.Find(id);
+            if (!Authenticate())
+                return null;
+
+            User usr = context.Users.Find(id);
+            usr.GetLinkedInformations();
+            return usr;
         }
 
         // POST api/users
@@ -60,8 +86,11 @@ namespace API.Controllers
         [HttpDelete("{id}")]
         public bool Delete(int id)
         {
+            if (!Authenticate())
+                return false;
+
             User user = context.Users.Find(id);
-            if (user != null)
+            if (user != null && user.Id != int.Parse(Request.Headers["userId"]))
             {
                 context.Users.Remove(user);
                 context.SaveChanges();
